slothjs: Overhaul of interpretation to accomodate for mutable places.
The lost paradise of pure values
================================

The previous version of the interpreter was very simple and
Haskell-like: it did not understood references, there was no way to
share references on the heap, the storage type for "objects" was JSON
internally. Everything was a value, nothing had identity, even arrays
and objects.

The heap, references and places
===============================

Now `JSValue` can be an object with references to other objects. This
requires some kind of multiple references to the same memory. I did not
want to overstretch Rust reference counting and decided to go with a
simple Vec-as-the-heap that contains all `JSValue`s that have identity.
Some values still can be ephemeral (e.g. literals, temporary valus in
expressions, etc).

The `Heap` type is a simple wrapper for a vector of `JSValue`. A
`JSValue` can be a primitive value or a `JSObject` with properies with
references to other `JSValue`s. Serializing an object to JSON now
requires the heap to be present as well, to read all the nested property
objects.

A mutable place on the heap is referenced by a `JSRef`, which is a thin
wrapper over an index in the heap vector. It is not entangled with Rust
references in any way, but again requires presence of the heap to
dereference the value.

An interesting catch is that mutable places are not always `JSRef`s.
E.g. an AssignmentExpression might have a left-hand MemberExpression.
A MemberExpression on its own does not know if it should create the
reference:

    > a = {}
    > a.one      # MemberExpression does not create a `JSRef`
    undefined
    > a.one = 1  # MemberExpression creates a `JSRef` for `one`

This requires the abstraction of `Place`: it can be a `JSRef` or an
instruction for future creation of `JSRef` when needed (in this case,
`Place::ObjectMember()`, which holds a `JSRef` to `a` and the identifier
`one`) and is reified later by AssignmentExpression.

TODO
====

- Garbage collection;
- Arrays as objects with `__proto__` equal to `Array.prototype`; maybe
optimized by a BTreeMap storage.