slothjs: getting the main data types right.
Initial mistakes
================

When I just started, it was fine to assume that the Heap contains `JSValue`s
directly, be it a `Number`, an `Undefined`, or even a `JSObject`. When I
started paying attention to modifiable places, i.e. object properies, it
became clear that my initial model in which JSRef referenced a JSValue
does not correspond to how Javascript really works. E.g. I had to
introduce different `Interpreted::Ref` and `Interpreted::Value`, even
though a reference is trivially assignable and copyable and is a value.
Then, when I read more about Javascript objects, it became clear that it
does not make much sense to put primitive values on the heap: only objects
should be there. Interpreted::Ref could not express everything that
Interpreted::Member could, and it complicated things without a good
reason.

The new data model
==================

Now the `Heap` is a vector of `JSObject`s, properties of which contains
`JSValue`. JSValue can now be either a primitive or a JSValue::Ref to
another object.

       JSRef(0)     JSRef(1)     JSRef(2)     JSRef(3)        ...
          v             v            v           v
    |------------|------------|------------|------------|---- ...
       JSObject     JSObject     JSObject     JSObject        ...

Each `JSObject` has:

- the `proto` field, which is a JSRef (a heap index) to the prototype
  object of this one (or `null`, which is `Heap::NULL`, the 0th reference).

- the `value: ObjectValue` field, which encapsulates either a primitive value
  for `Number`/`Boolean`/`String` or some internal structure like `Closure`,
  `VMCall` (a native function), `JSArray`, etc. `.value` is not directly
  accessible from Javascript anymore (before this I used a non-configurable,
  non-writable, non-enumerable property called `[[value]]`).

  Closures and native functions now must be wrapped in an object. This
  removed the headache of representing internal values in properties:
  JSObject.value is not a Javascript-visible field anymore.

  Properties are now simply a Content::Value or a Content::Accessor (to
  be done), not a mishmash of everything.

- and last but not least, the `properties` hashtable that contains
  string keys pointing to `JSValue`s. A JSValue can be a JSValue::Ref
  pointing to another object on the heap.

Builin objects
==============

Some standard prototype objects now live at fixed positions on the Heap:
e.g. Heap::OBJECT_PROTO, Heap::ARRAY_PROTO, Heap::FUNCTION_PROTO, etc.
This cleared the mess of src/builtin/mod.rs initialization considerably
and removed the need for expensive lookups when constructing new
instances of these.

`null` and `undefined`
=====================

I've cleared my initial confusion: `null` is not a primitive value,
but `undefined` is. `null` is a special kind of object, roughly
corresponding to a NULL pointer in C++. In this implementation, the 0th
object of the heap is now reserved to designate `null`.

Conclusions
===========

The refactoring was expansive, but mostly mechanical and the test suite
helped well. The results:

    $ git diff master --stat
    ...
    10 files changed, 510 insertions(+), 804 deletions(-)

I am very satisfied with the delta of minus 300 lines of code!