slothjs: change TryFrom to custom ParseFrom with a ParserContext
Stateless TryFrom<JSON> has served well, but it's time for a more
complicated solution.

This is preparation for better variable analysis: in order to support
variable hoisting and computing free variables of a closure (and block
scoped variables sometime in the future), it's necessary to accumulate
mentioned and declared variables in the parser.