slothjs: hoist function closures, not just function names
Closures are now created when entering a scope, in order to be
accessible in that scope before textual declaraion. In other words,
hoisted functions are now callable before their declaration is reached
by the interpreter.

This also marks another milestone: Esprima can be fully parsed and
loaded! Now we just need to iron out all of the runtime bugs of its
actual execution.