slothjs: String: constructor, valueOf(), toString(); indexing
String characters are now accessed very slowly via sequential access.

I had a bad lifetime conundrum with `JSObject::get_value()` not being
able to return `Option<&JSValue>` as a reference because of new addition
of string indexing: string characters do not live on the heap and trying
to put them there temporarily would suddenly change a lot of code.
Fortunately, `JSValue` is a very lightweight type, so it makes more
sense to return a new copy (maybe with an owning JSValue::String()
inside) than try to return a reference to it.