slothjs: SimpleCopyingGC draft
SimpleCopyingGC takes an existing heap vector, uses everything below
Heap::USERSTART as roots and visits everything from roots recursively,
relocating live objects into a new vector, `SimpleCopyingGC::objects`,
and leaving so called _forwarder_ objects behind on the heap, which
contain a JSRef in the new heap.

After all live objects are visited, the newly relocated objects are
patched to replace JSRefs on the old heap with JSRefs on the new heap
they now forward to. After that the new object vector becomes the heap
memory.

One huge problem of this approach is that any off-heap JSRef will be
left invalid, and there is a bunch of them in the code hiding in local
variables. Recursive structure of interpretation makes it difficult to
avoid local JSRefs surviving .interpret() calls.