sljs: WIP for GC-enabled design
As discovered previsously with `SimpleCopyingGC`, all roots must be
known during garbage collection. Putting any `JSRef` (and therefore any
`JSValue`/`JSObject`) on a native stack/heap makes this root unknown
to a GC.

So, the requirements for GC-enabled memory management are:

- a runtime MUST know all the roots somehow (not the case previously)
- if a moving GC is used, there must be a way to update any reference,
  external or internal to the heap.
- it's nice to have the least amount of pointer indirection possible.
  Tighly packed structures with good cache locality are a plus.

I've been thinking about two possible ways out.

1. Expose unleakable references to the outside code
===================================================

That means ending the `JSRef` free-for-all anarchy: they must not be
available outside of a well-controlled scope and it must be impossible
to copy any references/non-primitive values out of it.

This seems to be possible with a combination of borrow checking and
(gasp) continuation-passing-like style:

    pub struct Heap { ... }

    pub struct Handle { /* opaque fields */ }

    impl Heap {
        fn with_handle<F, T>(&mut self, action: F) -> T
            where F: FnMut(&mut Heap, &Handle) -> T
        {
            // Make a new handle, save it in the heap.
            // Handle can encapsulate values/references and provide
            // methods to extract primitive values or manipulate
            // objects behind references safely.
            self.handles.push(JSValue::Undefined);
            // Now GC knows about all `heap.handles` in use at any
            // point.
            let handle_index = self.handles.len() - 1;
            let handle: Handle = Handle(handle_index);

            // A Handle reference will have a restricted lifetime inside
            // `action`, which makes it impossible to copy it outside.
            // Since the reference is immutable and opaque for outside
            // code, it's not possible to copy/move out the value out
            // of the `action` scope, nor make the reference outlive
            // it, nor delete a Handle.
            // Heap can dereference handle_index to real values or
            // references.
            let result = action(self, &handle)

            // Clean it up
            self.handles.pop();
            result
        }
    }

    // e.g. implementing addition:
    impl Interpretable for BinaryExpression {
        fn evaluate(&self, result: &Handle, heap: &mut Heap) -> JSResult<()> {
            heap.evaluate(self.lhs, result)?;
            let lval = result.numberify();
            heap.evaluate(self.rhs, result)?;
            let rval = result.numberify();
            match self.binop {
                BinOp::Add => result.put_value(lval + rval),
                ...
            }
            Ok(())
        }
    }

    impl Interpretable for ObjectExpression {
        fn evaluate(&self, result: &Handle, heap: &mut Heap) -> JSResult<()> {
            heap.alloc(/*into*/ result);
            for (key, valexpr) in self.0.iter() {
                let keyname = ...;
                heap.with_local(|heap, value: &Handle| {
                    heap.evaluate(valexpr, /*into*/ value)?;
                    result.set_property(keyname, /*from*/ value)?;
                })?;
            }
            Ok(())
        }
    }

`&Handle` works like a fixed "variable" for external code that can
provide primitive values, be manipulated like an object via methods,
be filled with another value/reference.

Everything must be done via an extensive collection of methods
on opaque `Handle`s. This is not exactly the easiest way to program,
but a fair price to pay to have provably unleakable references.

This approach is still unsafe when two different heaps are used: nothing
prevents one heap from using Handles of another heap and getting very
confused.

2. Account for all external references
======================================

The continuation-passing style can be unpleasant. Could something more
familiar be provided to the user of the library, even at a cost at
runtime?

One way to do it is to have a "gateway" structure bundled with a Heap
that creates a layer of indirection for external references and gives
a place to adjust internal references when objects are moved on the
Heap.

2.1 External OwnRefs with Atomic<HIndex>
========================================

JSRef can access the heap via intermediate Rust-allocated `OwnRef`
structs that are Rc-owned by `JSRef`s and accessible from the `Heap`
via `Weak<OwnRef>` at the same time through

    heap.references: HashMap<HIndex, Weak<OwnRef>>

 that maps internal `HIndex`es to possibly living `OwnRef`s.  Keys of
`heap.references` contain all living roots.

An `OwnRef` also keep the same `Atomic<HIndex>` inside. Why `Atomic`?
`heap` can only access `OwnRef` via an upgraded `Rc<OwnRef>`, which
guarantees that `Rc::get_mut()` is impossible due to `.strong_count()`
being more than 1. `Rc::get()` is still possible in this situation
and `Atomic::store(&self, new_hindex)` provides an escape hatch to
safely update the value.

When a `OwnRef` is `Drop`ped, it will be lingering as a value in
`heap.references` until a garbage collection cycle looks at it and
sees 0 in `ownref.strong_count()`, then it can be cleaned up. If not,
`HIndex` can be updated inside the `OwnRef` to the new index and the
key of `heap.references` can be adjusted as well.

2.2. A compomise: keep externally accessible HIndexes the same
==============================================================

What if GC only knows which objects have living external references and
which ones do not have? In this case, a compomise is possible:
externally-accessible objects should be kept at their `HIndex`es, but
all other objects can be moved safely.

One way to know about external references is to keep a `Atomic` counter,
`n_ext`, for external references in each `Object` on the `Heap`.
Whenever a `Heap` produces a new reference, its object has its `n_ext`
incremented, whenever an external reference is dropped, it's
decremented.

A GC can look at `object.n_ext.load()` and if it's zero, `object` can be
moved or (in absence of internal references to it as well) garbage
collected. If it's non-zero, the object should be kept at this `HIndex`.
Free objects with lower indexes can be added to a "free list" for
allocation.

An external `JSRef` must also know its `Heap` via `Weak<Heap>`.
The `Atomic`ity of `n_ext` counters is useful to be able to update them
via an `Rc::upgrade`d immutable references in `<JSRef as Drop>::drop()`.

Who owns the `Heap`? There must be an external struct, say, `Runtime`
with a carefully guarded singleton `Rc<Heap>` in it.

Checking the provenance of `JSRef` becomes easy with an accessible
`Weak<Heap>`: now a heap can check if a `JSRef` refers to it before
trusting internal indexes in it.

A downside of this approach is that the owning `Rc<Heap>` must lock
itself into a `&mut Heap` at runtime on every mutable operation with the
Heap. It can also fail to do so, e.g. whenever a `JSRef` has a temporary
`Rc<Heap>` alive at the same time. A safety of this approach is not
completely clear at the moment.