sljs v0.2.0: `Realm::with` works; abandon opaque `Ref`s for now.
While I enjoyed the challenge of getting const arrays working, I feel
that I should abandon the experiment with opaque scoped `Ref`s.

The original problem was in tracking all the roots, including those in
external references to objects on a `Heap`, free-floating on Rust
stack/heap. It is always possible to solve the problem with a level of
runtime indirection: external references to `JSObject`s being just
updatable wrappers for internal indexes in their `Heap`.

Still, I was wishing for a zero-cost, statically-verified solution. I
did not want to optimize for an easy, "natural" style of programming,
judging that zero-cost static abstractions would be worth it, if
achieved.

I hoped that Rust borrow checker could be used creatively to limit scopes
for local opaque references to objects, while providing a provably safe
API. The idea was to narrow the API to `Realm`s (hiding and owning
`Heap`s via `Rc<RefCell<Heap>>`), with `Ref`s available in strictly
controlled scopes like this:

    Realm::with([With::New, With::Var("x")], |[obj, x]| {
        // use methods of `obj` to manage the object
    })?;

Internally, a `Ref` would encapsulate a `LRef`, an index of a local
`InnerRef` in `heap.refstack`, where an actual internal index of
a `JSObject` (`HRef`) would be stored, making it available for a GC to
change `HRef`s freely.

This already creates a couple of indirections and vaguely-named `*Ref`s.
With proper discipline inside `create::heap` this complex solution
can be made work, but this is obviously not a zero-cost solution.

Conclusions:

- this style of coding does not actually produce much zero-cost static
  guarantees that I hoped for: there's a dynamic indirection and lots of
  `RefCell::borrow_mut()`. That's obviously the problem mostly related
  to having multiple `Realm`s, and not limited scopes.
  Managing multiple `Realm`s efficiently needs more thought.

- the `::with()` closure parameters can be either `&Ref`, or `Pin<&mut Ref>`.
  The former option feels a bit too promiscuous and easy to create
  borrow panics when used without diligence. The second option is more
  restricting (in a good way), but hideous.