sljs: optimize Esprima loading time 5x.
A stray `.clone()` bites again: profiling showed that 31% of the time is
spent in `serde_json::Value::clone()` and 43% of the time in
`core::ptr::drop_in_place<ParseError>`.

It turned out that `<JSON as SourceNode>::get_location()` was indeed
horribly wrong: it tried to look for `.start` and `.end` in the node
itself, instead of `.loc.start` and `.loc.end`. Obviously, this was
failing, and `<JSON as SourceNode>::to_error()` was dutifully and
diligently cloning subtrees all over the place, eating >80% of Esprima
loading time.

Now all of Esprima JSON AST can be loaded in less than 80ms instead of
~400ms.